2025/adk-agentengine-basic/step04/requirements.txt
google-cloud-aiplatform[adk,agent_engines]==1.106.0
cloudpickle==3.1.1
uv==0.8.4
pydantic==2.11.7
